[frontend] Move export creation button to StixCoreObjectsExportsContent (#13303)
